requirements.txt
# Core web app
Flask>=3.0,<4

# PDF parsing and layout-aware chunking
unstructured[pdf]>=0.12
pymupdf>=1.23
Pillow>=10

# Embeddings and retrieval
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24

# Torch stack (CPU by default via PyPI). If you need CUDA, install from pytorch.org instructions.
torch
torchvision
torchaudio

# Optional OCR (system tesseract required). Uncomment if enabling OCR.
# pytesseract
# easyocr


